refactor(backend): split adapter.py into Mixin modules (Issue #117)
